started to add scripting

+++ Source/Voxel/CSG.cpp
@@ -14,7 +14,7 @@ void CSG::apply(const std::vector<bool>& a, std::vector<bool>& b, std::function<
   assert(a.size() == b.size());
   assert(a.size() % 8 == 0);
 #pragma omp parallel for
-  for (size_t i = 0; i < a.size(); i+=8) {
+  for (long long i = 0; i < a.size(); i+=8) {
     for (int sub = 0; sub < 8; sub++) { //std::vector<bool> is a special flower in terms of parallelism
       b[i + sub] = fun(b[i + sub], a[i + sub]);
     }
